api-gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pika==1.3.2
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
deepface==0.0.79
retina-face==0.0.13
pgvector
aio-pika
tensorflow==2.15.0
tf-keras
python-jose[cryptography]
passlib[bcrypt]